@@ -1,26 +1,23 @@
-<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Aspire.AppHost.Sdk/13.0.2">
 
-  <Sdk Name="Aspire.Hosting.Sdk" Version="8.2.2" />
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net10.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <UserSecretsId>39ce7757-c12c-4e0b-b7e5-51c82fb69785</UserSecretsId>
+    </PropertyGroup>
 
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <UserSecretsId>39ce7757-c12c-4e0b-b7e5-51c82fb69785</UserSecretsId>
-  </PropertyGroup>
+    <ItemGroup>
+        <PackageReference Include="Aspire.Hosting.Azure.CosmosDB" Version="13.0.2"/>
+        <PackageReference Include="Aspire.Hosting.PostgreSQL" Version="13.0.2"/>
+        <PackageReference Include="Aspire.Hosting.RabbitMQ" Version="13.0.2"/>
+        <PackageReference Include="Aspire.Hosting.Redis" Version="13.0.2"/>
+    </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Aspire.Hosting.AppHost" Version="13.0.2" />
-    <PackageReference Include="Aspire.Hosting.Azure.CosmosDB" Version="13.0.2" />
-    <PackageReference Include="Aspire.Hosting.PostgreSQL" Version="13.0.2" />
-    <PackageReference Include="Aspire.Hosting.RabbitMQ" Version="13.0.2" />
-    <PackageReference Include="Aspire.Hosting.Redis" Version="13.0.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Dometrain.Monolith.Api\Dometrain.Monolith.Api.csproj" />
-    <ProjectReference Include="..\Dometrain.Aspire.ServiceDefaults\Dometrain.Aspire.ServiceDefaults.csproj" />
-  </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\Dometrain.Monolith.Api\Dometrain.Monolith.Api.csproj"/>
+        <ProjectReference Include="..\Dometrain.Aspire.ServiceDefaults\Dometrain.Aspire.ServiceDefaults.csproj"/>
+    </ItemGroup>
 
 </Project>
