@@ -1,6 +1,9 @@
+#region
+
 using System.Text.Json;
-using Dometrain.Monolith.Api.Database;
 using StackExchange.Redis;
+
+#endregion
 
 namespace Dometrain.Monolith.Api.Courses;
 
@@ -18,10 +21,7 @@ public class CachedCourseRepository : ICourseRepository
     public async Task<Course?> CreateAsync(Course course)
     {
         var created = await _courseRepository.CreateAsync(course);
-        if (created is null)
-        {
-            return null;
-        }
+        if (created is null) return null;
 
         var db = _multiplexer.GetDatabase();
         var serializedCourse = JsonSerializer.Serialize(course);
@@ -30,7 +30,7 @@ public class CachedCourseRepository : ICourseRepository
             new($"course:id:{course.Id.ToString()}", serializedCourse),
             new($"course:slug:{course.Slug}", course.Id.ToString())
         };
-        
+
         await db.StringSetAsync(batch);
         return created;
     }
@@ -39,18 +39,13 @@ public class CachedCourseRepository : ICourseRepository
     {
         var db = _multiplexer.GetDatabase();
         var cachedCourse = await db.StringGetAsync($"course:id:{id}");
-        if (!cachedCourse.IsNull)
-        {
-            return JsonSerializer.Deserialize<Course>(cachedCourse.ToString());
-        }
-    
+        if (!cachedCourse.IsNull) return JsonSerializer.Deserialize<Course>(cachedCourse.ToString());
+
         var course = await _courseRepository.GetByIdAsync(id);
-        if (course is null)
-        {
-            return course;
-        }
+        if (course is null) return course;
+
         var serializedCourse = JsonSerializer.Serialize(course);
-        
+
         var batch = new KeyValuePair<RedisKey, RedisValue>[]
         {
             new($"course:id:{course.Id.ToString()}", serializedCourse),
@@ -64,17 +59,12 @@ public class CachedCourseRepository : ICourseRepository
     {
         var db = _multiplexer.GetDatabase();
         var cachedCourseKey = await db.StringGetAsync($"course:slug:{slug}");
-    
-        if (!cachedCourseKey.IsNull)
-        {
-            return await GetByIdAsync(Guid.Parse(cachedCourseKey.ToString()));
-        }
-    
+
+        if (!cachedCourseKey.IsNull) return await GetByIdAsync(Guid.Parse(cachedCourseKey.ToString()));
+
         var course = await _courseRepository.GetBySlugAsync(slug);
-        if (course is null)
-        {
-            return course;
-        }
+        if (course is null) return course;
+
         var serializedCourse = JsonSerializer.Serialize(course);
         var batch = new KeyValuePair<RedisKey, RedisValue>[]
         {
@@ -87,17 +77,14 @@ public class CachedCourseRepository : ICourseRepository
 
     public Task<IEnumerable<Course>> GetAllAsync(string nameFilter, int pageNumber, int pageSize)
     {
-        return _courseRepository.GetAllAsync(nameFilter, pageNumber, pageSize);   
+        return _courseRepository.GetAllAsync(nameFilter, pageNumber, pageSize);
     }
 
     public async Task<Course?> UpdateAsync(Course course)
     {
         var updated = await _courseRepository.UpdateAsync(course);
-        if (updated is null)
-        {
-            return updated;
-        }
-    
+        if (updated is null) return updated;
+
         var db = _multiplexer.GetDatabase();
         var serializedCourse = JsonSerializer.Serialize(course);
         var batch = new KeyValuePair<RedisKey, RedisValue>[]
@@ -113,19 +100,14 @@ public class CachedCourseRepository : ICourseRepository
     {
         var deleted = await _courseRepository.DeleteAsync(id);
 
-        if (!deleted)
-        {
-            return deleted;
-        }
-        
+        if (!deleted) return deleted;
+
         var db = _multiplexer.GetDatabase();
         var cachedCourseString = await db.StringGetAsync($"course:id:{id}");
-        if (cachedCourseString.IsNull)
-        {
-            return deleted;
-        }
-        var course = JsonSerializer.Deserialize<Course>(cachedCourseString!)!;
-        var deletedCache = await db.KeyDeleteAsync([$"course:id:{id}",$"course:slug:{course.Slug}"]);
+        if (cachedCourseString.IsNull) return deleted;
+
+        var course = JsonSerializer.Deserialize<Course>(cachedCourseString.ToString())!;
+        var deletedCache = await db.KeyDeleteAsync([$"course:id:{id}", $"course:slug:{course.Slug}"]);
         return deletedCache > 0;
     }
 }
