@@ -1,4 +1,8 @@
+#region
+
 using System.Text.RegularExpressions;
+
+#endregion
 
 namespace Dometrain.Monolith.Api.Courses;
 
@@ -7,13 +11,13 @@ public partial class Course
     public required Guid Id { get; init; }
 
     public required string Name { get; set; }
-    
+
     public required string Description { get; set; }
 
     public string Slug => GenerateSlug();
 
     public required string Author { get; set; }
-    
+
     private string GenerateSlug()
     {
         return SlugRegex().Replace(Name, string.Empty)
