@@ -1,5 +1,9 @@
+#region
+
 using System.Data;
 using Npgsql;
+
+#endregion
 
 namespace Dometrain.Monolith.Api.Database;
 
@@ -8,17 +12,10 @@ public interface IDbConnectionFactory
     Task<IDbConnection> CreateConnectionAsync(CancellationToken token = default);
 }
 
-public class NpgsqlConnectionFactory : IDbConnectionFactory
+public class NpgsqlConnectionFactory(NpgsqlDataSource dataSource) : IDbConnectionFactory
 {
-    private readonly NpgsqlDataSource _dataSource;
-
-    public NpgsqlConnectionFactory(NpgsqlDataSource dataSource)
-    {
-        _dataSource = dataSource;
-    }
-
     public async Task<IDbConnection> CreateConnectionAsync(CancellationToken token = default)
     {
-        return await _dataSource.OpenConnectionAsync(token);
+        return await dataSource.OpenConnectionAsync(token);
     }
 }
