@@ -1,5 +1,9 @@
+#region
+
 using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
+
+#endregion
 
 namespace Dometrain.Monolith.Api.Identity;
 
@@ -20,20 +24,17 @@ public class AdminAuthRequirement : IAuthorizationHandler, IAuthorizationRequire
             context.Succeed(this);
             return Task.CompletedTask;
         }
-        
+
         var httpContext = context.Resource as HttpContext;
-        if (httpContext is null)
-        {
-            return Task.CompletedTask;
-        }
-        
+        if (httpContext is null) return Task.CompletedTask;
+
         if (!httpContext.Request.Headers.TryGetValue("x-api-key", out
                 var extractedApiKey))
         {
             context.Fail();
             return Task.CompletedTask;
         }
-        
+
         if (_apiKey != extractedApiKey)
         {
             context.Fail();
