@@ -1,8 +1,12 @@
+#region
+
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
+
+#endregion
 
 namespace Dometrain.Monolith.Api.Identity;
 
@@ -39,16 +43,17 @@ public class IdentityService : IIdentityService
             var claim = new Claim("is_admin", "true", ClaimValueTypes.Boolean);
             claims.Add(claim);
         }
-        
+
         var tokenDescriptor = new SecurityTokenDescriptor
         {
             Subject = new ClaimsIdentity(claims),
             Expires = DateTime.UtcNow.Add(_identitySettings.Value.Lifetime),
             Issuer = _identitySettings.Value.Issuer,
             Audience = _identitySettings.Value.Audience,
-            SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
+            SigningCredentials =
+                new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
         };
-        
+
         var token = tokenHandler.CreateToken(tokenDescriptor);
 
         return tokenHandler.WriteToken(token);
