@@ -6,14 +6,11 @@ public static class IdentityExtensions
     {
         var userId = context.User.Claims.SingleOrDefault(x => x.Type == "user_id");
 
-        if (Guid.TryParse(userId?.Value, out var parsedId))
-        {
-            return parsedId;
-        }
+        if (Guid.TryParse(userId?.Value, out var parsedId)) return parsedId;
 
         return null;
     }
-    
+
     public static bool IsAdmin(this HttpContext context)
     {
         var isAdmin = context.User.Claims.SingleOrDefault(x => x is { Type: "is_admin", Value: "true" });
