@@ -1,5 +1,9 @@
+#region
+
 using Dometrain.Monolith.Api.Enrollments;
 using Dometrain.Monolith.Api.Students;
+
+#endregion
 
 namespace Dometrain.Monolith.Api.Orders;
 
@@ -14,11 +18,12 @@ public interface IOrderService
 
 public class OrderService : IOrderService
 {
-    private readonly IOrderRepository _orderRepository;
     private readonly IEnrollmentService _enrollmentService;
+    private readonly IOrderRepository _orderRepository;
     private readonly IStudentRepository _studentRepository;
-    
-    public OrderService(IOrderRepository orderRepository, IStudentRepository studentRepository, IEnrollmentService enrollmentService)
+
+    public OrderService(IOrderRepository orderRepository, IStudentRepository studentRepository,
+        IEnrollmentService enrollmentService)
     {
         _orderRepository = orderRepository;
         _studentRepository = studentRepository;
@@ -38,24 +43,16 @@ public class OrderService : IOrderService
     public async Task<Order?> PlaceAsync(Guid studentId, IEnumerable<Guid> courseIds)
     {
         var student = await _studentRepository.GetByIdAsync(studentId);
-        if (student is null)
-        {
-            return null;
-        }
-        
+        if (student is null) return null;
+
         var order = await _orderRepository.PlaceAsync(studentId, courseIds);
 
-        if (order is null)
-        {
-            return null;
-        }
+        if (order is null) return null;
 
         var enrollments = await _enrollmentService.GetStudentEnrollmentsAsync(studentId);
-        
+
         foreach (var courseId in order.CourseIds.Where(x => !enrollments!.CourseIds.Contains(x)))
-        {
             await _enrollmentService.EnrollToCourseAsync(studentId, courseId);
-        }
 
         return order;
     }
