@@ -1,15 +1,19 @@
+#region
+
 using System.Text.Json;
 using StackExchange.Redis;
+
+#endregion
 
 namespace Dometrain.Monolith.Api.ShoppingCarts;
 
 public class CachedShoppingCartRepository : IShoppingCartRepository
 {
-    private readonly IShoppingCartRepository _shoppingCartRepository;
     private readonly IConnectionMultiplexer _connectionMultiplexer;
+    private readonly IShoppingCartRepository _shoppingCartRepository;
 
     public CachedShoppingCartRepository(
-        IShoppingCartRepository shoppingCartRepository, 
+        IShoppingCartRepository shoppingCartRepository,
         IConnectionMultiplexer connectionMultiplexer)
     {
         _shoppingCartRepository = shoppingCartRepository;
@@ -20,10 +24,8 @@ public class CachedShoppingCartRepository : IShoppingCartRepository
     {
         var added = await _shoppingCartRepository.AddCourseAsync(studentId, courseId);
 
-        if (!added)
-        {
-            return added;
-        }
+        if (!added) return added;
+
         var db = _connectionMultiplexer.GetDatabase();
         await db.KeyDeleteAsync($"cart:id:{studentId}");
         return added;
@@ -33,11 +35,8 @@ public class CachedShoppingCartRepository : IShoppingCartRepository
     {
         var db = _connectionMultiplexer.GetDatabase();
         var cachedCartString = await db.StringGetAsync($"cart:id:{studentId}");
-        if (!cachedCartString.IsNull)
-        {
-            return JsonSerializer.Deserialize<ShoppingCart>(cachedCartString.ToString());
-        }
-        
+        if (!cachedCartString.IsNull) return JsonSerializer.Deserialize<ShoppingCart>(cachedCartString.ToString());
+
         var cart = await _shoppingCartRepository.GetByIdAsync(studentId);
         await db.StringSetAsync($"cart:id:{studentId}", JsonSerializer.Serialize(cart));
         return cart;
@@ -46,11 +45,8 @@ public class CachedShoppingCartRepository : IShoppingCartRepository
     public async Task<bool> RemoveItemAsync(Guid studentId, Guid courseId)
     {
         var removed = await _shoppingCartRepository.RemoveItemAsync(studentId, courseId);
-        if (!removed)
-        {
-            return removed;
-        }
-        
+        if (!removed) return removed;
+
         var db = _connectionMultiplexer.GetDatabase();
         await db.KeyDeleteAsync($"cart:id:{studentId}");
         return removed;
@@ -59,11 +55,8 @@ public class CachedShoppingCartRepository : IShoppingCartRepository
     public async Task<bool> ClearAsync(Guid studentId)
     {
         var cleared = await _shoppingCartRepository.ClearAsync(studentId);
-        if (!cleared)
-        {
-            return cleared;
-        }
-        
+        if (!cleared) return cleared;
+
         var db = _connectionMultiplexer.GetDatabase();
         await db.KeyDeleteAsync($"cart:id:{studentId}");
         return cleared;
