@@ -1,4 +1,8 @@
+#region
+
 using Newtonsoft.Json;
+
+#endregion
 
 namespace Dometrain.Monolith.Api.ShoppingCarts;
 
@@ -7,9 +11,8 @@ public class ShoppingCart
     [JsonProperty("pk")]
     [System.Text.Json.Serialization.JsonIgnore]
     public string Pk => StudentId.ToString();
-    
-    [JsonProperty("id")]
-    public required Guid StudentId { get; set; }
+
+    [JsonProperty("id")] public required Guid StudentId { get; set; }
 
     public List<Guid> CourseIds { get; set; } = [];
 }
