@@ -1,4 +1,8 @@
+#region
+
 using Dometrain.Monolith.Api.Identity;
+
+#endregion
 
 namespace Dometrain.Monolith.Api.ShoppingCarts;
 
@@ -10,21 +14,23 @@ public static class ShoppingCartEndpoints
         var cart = await shoppingCartService.GetByIdAsync(studentId.Value);
         return cart is null ? Results.NotFound() : Results.Ok(cart);
     }
-    
-    public static async Task<IResult> AddCourse(Guid courseId, IShoppingCartService shoppingCartService, HttpContext httpContext)
+
+    public static async Task<IResult> AddCourse(Guid courseId, IShoppingCartService shoppingCartService,
+        HttpContext httpContext)
     {
         var studentId = httpContext.GetUserId()!;
         var cart = await shoppingCartService.AddCourseAsync(studentId.Value, courseId);
         return cart is null ? Results.NotFound() : Results.Ok(cart);
     }
-    
-    public static async Task<IResult> RemoveCourse(Guid courseId, IShoppingCartService shoppingCartService, HttpContext httpContext)
+
+    public static async Task<IResult> RemoveCourse(Guid courseId, IShoppingCartService shoppingCartService,
+        HttpContext httpContext)
     {
         var studentId = httpContext.GetUserId()!;
         var cart = await shoppingCartService.RemoveItemAsync(studentId.Value, courseId);
         return cart is null ? Results.NotFound() : Results.Ok(cart);
     }
-    
+
     public static async Task<IResult> ClearCart(IShoppingCartService shoppingCartService, HttpContext httpContext)
     {
         var studentId = httpContext.GetUserId()!;
