@@ -1,5 +1,9 @@
+#region
+
 using Dometrain.Monolith.Api.Courses;
 using Dometrain.Monolith.Api.Students;
+
+#endregion
 
 namespace Dometrain.Monolith.Api.ShoppingCarts;
 
@@ -8,19 +12,20 @@ public interface IShoppingCartService
     Task<ShoppingCart?> AddCourseAsync(Guid studentId, Guid courseId);
 
     Task<ShoppingCart?> GetByIdAsync(Guid studentId);
-    
+
     Task<ShoppingCart?> RemoveItemAsync(Guid studentId, Guid courseId);
-    
+
     Task<ShoppingCart?> ClearAsync(Guid studentId);
 }
 
 public class ShoppingCartService : IShoppingCartService
 {
+    private readonly ICourseRepository _courseRepository;
     private readonly IShoppingCartRepository _shoppingCartRepository;
     private readonly IStudentRepository _studentRepository;
-    private readonly ICourseRepository _courseRepository;
 
-    public ShoppingCartService(IShoppingCartRepository shoppingCartRepository, IStudentRepository studentRepository, ICourseRepository courseRepository)
+    public ShoppingCartService(IShoppingCartRepository shoppingCartRepository, IStudentRepository studentRepository,
+        ICourseRepository courseRepository)
     {
         _shoppingCartRepository = shoppingCartRepository;
         _studentRepository = studentRepository;
@@ -30,17 +35,11 @@ public class ShoppingCartService : IShoppingCartService
     public async Task<ShoppingCart?> AddCourseAsync(Guid studentId, Guid courseId)
     {
         var student = await _studentRepository.GetByIdAsync(studentId);
-        if (student is null)
-        {
-            return null;
-        }
+        if (student is null) return null;
 
         var course = await _courseRepository.GetByIdAsync(courseId);
-        if (course is null)
-        {
-            return null;
-        }
-        
+        if (course is null) return null;
+
         await _shoppingCartRepository.AddCourseAsync(studentId, courseId);
         return await GetByIdAsync(studentId);
     }
@@ -53,17 +52,11 @@ public class ShoppingCartService : IShoppingCartService
     public async Task<ShoppingCart?> RemoveItemAsync(Guid studentId, Guid courseId)
     {
         var student = await _studentRepository.GetByIdAsync(studentId);
-        if (student is null)
-        {
-            return null;
-        }
-        
+        if (student is null) return null;
+
         var course = await _courseRepository.GetByIdAsync(courseId);
-        if (course is null)
-        {
-            return null;
-        }
-        
+        if (course is null) return null;
+
         await _shoppingCartRepository.RemoveItemAsync(studentId, courseId);
         return await GetByIdAsync(studentId);
     }
@@ -71,10 +64,7 @@ public class ShoppingCartService : IShoppingCartService
     public async Task<ShoppingCart?> ClearAsync(Guid studentId)
     {
         var student = await _studentRepository.GetByIdAsync(studentId);
-        if (student is null)
-        {
-            return null;
-        }
+        if (student is null) return null;
 
         await _shoppingCartRepository.ClearAsync(studentId);
         return new ShoppingCart
