@@ -5,6 +5,6 @@ public class Student
     public required Guid Id { get; init; }
 
     public required string Email { get; init; }
-    
+
     public required string FullName { get; init; }
 }
