@@ -1,18 +1,23 @@
+#region
+
 using Dometrain.Monolith.Api.Identity;
+
+#endregion
 
 namespace Dometrain.Monolith.Api.Students;
 
 public static class StudentEndpoints
 {
     public const string GetStudentEndpointName = "GetStudent";
-    
+
     public static async Task<IResult> Register(StudentRegistrationRequest request, IStudentService studentService)
     {
         var student = request.MapToStudent();
         var createdStudent = await studentService.CreateAsync(student, request.Password);
-        return TypedResults.CreatedAtRoute(createdStudent, GetStudentEndpointName, new { idOrEmail = createdStudent!.Id });
+        return TypedResults.CreatedAtRoute(createdStudent, GetStudentEndpointName,
+            new { idOrEmail = createdStudent!.Id });
     }
-    
+
     public static async Task<IResult> Get(string idOrEmail, IStudentService studentService)
     {
         var isId = Guid.TryParse(idOrEmail, out var id);
@@ -23,20 +28,18 @@ public static class StudentEndpoints
     public static async Task<IResult> GetMe(HttpContext httpContext, IStudentService studentService)
     {
         var userId = httpContext.GetUserId();
-        if (userId is null)
-        {
-            return Results.NotFound();
-        }
+        if (userId is null) return Results.NotFound();
+
         var student = await studentService.GetByIdAsync(userId.Value);
         return student is null ? Results.NotFound() : Results.Ok(student);
     }
-    
+
     public static async Task<IResult> GetAll(IStudentService studentService, int pageNumber = 1, int pageSize = 25)
     {
         var students = await studentService.GetAllAsync(pageNumber, pageSize);
         return Results.Ok(students);
     }
-    
+
     public static async Task<IResult> Delete(Guid id, IStudentService studentService)
     {
         var deleted = await studentService.DeleteAsync(id);
